requirements.txt
appdirs==1.4.4
arrow==0.17.0
asgiref==3.2.10
backcall==0.2.0
binaryornot==0.4.4
blessings==1.7
boto3==1.16.48
botocore==1.19.49
certifi==2018.8.24
cffi==1.14.4
chardet==3.0.4
click==7.1.2
colorama==0.4.4
cookiecutter==1.7.2
coreapi==2.3.3
coreschema==0.0.4
coverage==4.5.1
coveralls==1.5.1
cryptography==3.3.1
decorator==4.4.2
defusedxml==0.6.0
distlib==0.3.1
dj-database-url==0.5.0
Django==2.2.13
django-allauth==0.43.0
django-configurations==2.2
django-cors-headers==3.6.0
django-cors-middleware==1.2.0
django-extensions==1.7.1
django-filter==2.0.0
django-heroku==0.3.1
django-model-utils==4.1.1
django-nose==1.4.7
django-rest-auth==0.9.5
django-storages==1.11.1
django-unique-upload==0.2.1
djangorestframework==3.9.1
djangorestframework-jwt==1.11.0
docopt==0.6.2
drf-yasg==1.20.0
factory-boy==2.11.1
Faker==0.9.2
filelock==3.0.12
flake8==3.8.4
future==0.18.2
gunicorn==20.0.4
idna==2.7
inflection==0.5.1
ipdb==0.13.4
ipython==7.19.0
ipython-genutils==0.2.0
itypes==1.2.0
jedi==0.18.0
Jinja2==2.11.2
jinja2-time==0.2.0
jmespath==0.10.0
joblib==1.0.0
livereload==2.6.3
lunr==0.5.8
Markdown==3.0.1
MarkupSafe==1.1.1
mccabe==0.6.1
mkdocs==1.1.2
mock==4.0.3
newrelic==5.24.0.153
nltk==3.5
nose==1.3.7
nose-progressive==1.5.2
oauthlib==3.1.0
packaging==20.8
parso==0.8.1
pickleshare==0.7.5
poyo==0.5.0
prompt-toolkit==3.0.9
psycopg2==2.8.6
psycopg2-binary==2.8.6
pycodestyle==2.6.0
pycparser==2.20
pyflakes==2.2.0
Pygments==2.7.3
PyJWT==1.6.4
pyparsing==2.4.7
python-dateutil==2.7.3
python-slugify==4.0.1
python3-openid==3.2.0
pytz==2018.5
PyYAML==5.3.1
regex==2020.11.13
requests==2.20.0
requests-oauthlib==1.3.0
ruamel.yaml==0.16.12
s3transfer==0.3.3
six==1.10.0
sqlparse==0.4.1
text-unidecode==1.2
tornado==6.1
tqdm==4.55.1
traitlets==5.0.5
uritemplate==3.0.1
urllib3==1.24.2
virtualenv==20.2.2
wcwidth==0.2.5
whitenoise==5.2.0
